Rename MetadataApi file
